requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-redis==5.0.0
djangorestframework==3.12.4
pycodestyle==2.7.0
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
toml==0.10.2
